### What
This PR adds baseline profiles generator, startup metric benchmark to see the difference between startup timings with/out baseline profiles. The goal is to test the change in startup timings in nightly. 

The metrics to consider:
– [Google Play Vitals - Nightly - Slow Cold Start](https://play.google.com/console/u/0/developers/7083182635971239206/app/4975374341268119750/vitals/metrics/details?peersetKey=1%3A37fd48e7&metric=COLD_STARTUP_TIME&days=28)
– [Glean - cold_x_app_to_first_frame](https://dictionary.telemetry.mozilla.org/apps/fenix?page=1&search=perf.startup) - I think it's interesting to monitor other first frame metrics along with `cold_main_app_to_first_frame` as the benefits can be in other metrics as well since the startup path has common elements.


### Local test data 
#### Using BaselineProfilesStartupBenchmark on Pixel 7 - Android 14

| Benchmark                                        | Time to Initial Display (ms) |
| ------------------------------------------------ | ---------------------------- |
| BaselineProfilesStartupBenchmark_startupPartialWithBaselineProfiles | min: 590.2, median: 642.9, max: 673.4 |
| BaselineProfilesStartupBenchmark_startupNone     | min: 732.4, median: 759.1, max: 790.2 |

#### Using BaselineProfilesStartupBenchmark on Samsung Galaxy A50 - Android 11
| Benchmark                                                        | Time to Initial Display (ms)               |
| ---------------------------------------------------------------- | ----------------------------------------- |
| BaselineProfilesStartupBenchmark_startupPartialWithBaselineProfiles | min: 917.3, median: 1,036.5, max: 1,089.5 |
| BaselineProfilesStartupBenchmark_startupNone                     | min: 1,388.7, median: 1,447.2, max: 1,479.9 |


#### Using ColdStartupDurationTelemetry on Pixel 7 - Android 14

| Type                              | Measurements                                  |
| --------------------------------- | --------------------------------------------- |
| Without Baseline Profiles | 635 ms, 594 ms, 440 ms, 427 ms, 420 ms |
| With Baseline Profiles)   | 548 ms, 451 ms, 431 ms, 367 ms, 368 ms |


Note: many times `startupStateProvider.isColdStartForStartedActivity()` returns false even though the conditions for cold start are present. So I removed that check to get the data. Between each run, the app was removed from recents and then force closed.



### Pull Request checklist
<!-- Before submitting the PR, please address each item -->
- [ ] **Quality**: This PR builds and passes detekt/ktlint checks (A pre-push hook is recommended)
- [ ] **Tests**: This PR includes thorough tests or an explanation of why it does not
- [ ] **Changelog**: This PR includes [a changelog entry](https://github.com/mozilla-mobile/firefox-android/blob/main/docs/changelog.md) or does not need one
- [ ] **Accessibility**: The code in this PR follows [accessibility best practices](https://github.com/mozilla-mobile/firefox-android/blob/main/docs/shared/android/accessibility_guide.md) or does not include any user facing features

### After merge
- **Breaking Changes**: If this is a breaking Android Components change, please push a draft PR on [Reference Browser](https://github.com/mozilla-mobile/reference-browser) to address the breaking issues.

### To download an APK when reviewing a PR (after all CI tasks finished running):
1. Click on `Checks` at the top of the PR page.
2. Click on the `firefoxci-taskcluster` group on the left to expand all tasks.
3. Click on the `build-apk-{fenix,focus,klar}-debug` task you're interested in.
4. Click on `View task in Taskcluster` in the new `DETAILS` section.
5. The APK links should be on the right side of the screen, named for each CPU architecture.












### GitHub Automation
https://bugzilla.mozilla.org/show_bug.cgi?id=1807350